+++ packages/datadog_grpc_interceptor/lib/src/datadog_grpc_interceptor.dart
@@ -67,9 +67,9 @@ class DatadogGrpcInterceptor extends ClientInterceptor {
             rum.traceSampleRate / 100.0;
         if (tracingContext.sampled) {
           attributes[DatadogRumPlatformAttributeKey.traceID] =
-              tracingContext.traceId.asString(TraceIdRepresentation.decimal);
+              tracingContext.traceId.asString(TracingIdRepresentation.hex);
           attributes[DatadogRumPlatformAttributeKey.spanID] =
-              tracingContext.spanId.asString(TraceIdRepresentation.decimal);
+              tracingContext.spanId.asString(TracingIdRepresentation.decimal);
         }
 
         for (final tracingType in headerTypes) {